MVVM architecture
